Moved Notification Center and NSObject Sink to separate files
